requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
pydantic==2.5.0
openai==1.35.0
pytest==7.4.0
pytest-cov==4.1.0
httpx==0.25.0
python-dotenv==1.0.0
